+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.2
 - Fix listing duplicate node modules and their identifiers.
+- Fix adding only named imports for cached node module identifiers.
 
 ## 1.4.1
 - Add file name for exported identifiers.